pyproject.toml
[tool.poetry]
name = "stock-analysis"
version = "0.1.0"
description = ""
authors = ["Tony <tonytimoshe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
kafka-python = "^2.0.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
